Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
Pillow
numpy
mediapipe==0.10.9
